Simplify deleteCustomTagLogic
